Move files around
Moved the files around.
Accidentally placed the training Colab notebook in the wrong folder, moved it to the correct one.

Most of the files in the "Self-Driving Car Project" directory were not useful or critical to the project. They were mostly stuff I was playing around with while getting the project to work. They now live in the misc sub-folder.